+++ tests/helpers/mod.rs
@@ -1,3 +1,7 @@
+#![allow(unused_imports)]
+
+mod debug;
 mod writer;
 
+pub use debug::debug;
 pub use writer::Writer;